+++ src/main/java/ru/cbr/demorestservice/domain/repository/CastomRepository.java
@@ -1,0 +1,18 @@
+package ru.cbr.demorestservice.domain.repository;
+
+import com.querydsl.core.types.dsl.EntityPathBase;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.querydsl.QuerydslPredicateExecutor;
+import org.springframework.data.querydsl.binding.QuerydslBinderCustomizer;
+import org.springframework.data.querydsl.binding.QuerydslBindings;
+import org.springframework.data.repository.NoRepositoryBean;
+
+import java.io.Serializable;
+
+@NoRepositoryBean
+interface CustomRepository<T, P extends EntityPathBase<T>, ID extends Serializable>
+        extends JpaRepository<T, ID>, QuerydslPredicateExecutor<T>, QuerydslBinderCustomizer<P> {
+
+    @Override
+    void customize(QuerydslBindings bindings, P root);
+}
