+++ src/main/java/ru/cbr/demorestservice/domain/type/Ogrn.java
@@ -1,4 +1,4 @@
-package ru.cbr.demorestservice.domain.model;
+package ru.cbr.demorestservice.domain.type;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -12,16 +12,16 @@ import javax.validation.constraints.Size;
 import java.io.Serializable;
 
 
-@NoArgsConstructor
 @AllArgsConstructor
+@NoArgsConstructor
 @Audited
 @Getter
 @Setter
-public class CreditOrganizationRegNum implements Serializable {
+public class Ogrn implements Serializable {
 
     @Column
-    @Size(max = 4)
-    // с пробелами @Pattern(regexp = "/^(?!.*(\\d).*\\1)[1-4](?: *, *[1-4])*$/")
-    @Pattern(regexp = "/^(?!.*(\\d).*\\1)[1-4](?:,[1-4])*$/")
-    private String regNum;
+    @Size(min = 13, max = 13)
+    @Pattern(regexp = "/^(?!.*(\\d).*\\1)[13](?: *, *[13])*$/")
+    private String ogrn;
+
 }
